Add files via upload
This project required quite a bit of data cleaning. It was quite interesting to explore data in depth and choose the approach for data cleaning etc. After that, tried to fit a logistic regression model to the data. Every algorithm was written from scratch (without making use of advanced libraries like sklearn). The thought process is documented extensively!